+++ de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/ad/trx/processor/api/impl/TrxItemExecutorBuilder.java
@@ -35,6 +35,7 @@ import org.adempiere.ad.trx.processor.spi.ITrxItemChunkProcessor;
 import org.adempiere.ad.trx.processor.spi.ITrxItemProcessor;
 import org.adempiere.util.Check;
 import org.adempiere.util.Services;
+import org.compiere.util.Env;
 
 /* package */class TrxItemExecutorBuilder<IT, RT> implements ITrxItemExecutorBuilder<IT, RT>
 {
@@ -45,7 +46,7 @@ import org.adempiere.util.Services;
 	// Context
 	private ITrxItemProcessorContext _processorCtx;
 	private Properties _ctx = null;
-	private String _trxName = null;
+	private String _trxName = ITrx.TRXNAME_ThreadInherited;
 
 	private ITrxItemProcessor<IT, RT> _processor;
 	private ITrxItemExceptionHandler _exceptionHandler = ITrxItemProcessorExecutor.DEFAULT_ExceptionHandler;
@@ -113,9 +114,9 @@ import org.adempiere.util.Services;
 			return _processorCtx;
 		}
 
-		Check.assumeNotNull(_ctx, "ctx is set");
+		final Properties ctx = _ctx != null ? _ctx : Env.getCtx();
 		final ITrx trx = trxManager.getTrx(_trxName);
-		final ITrxItemProcessorContext processorCtx = executorService.createProcessorContext(_ctx, trx);
+		final ITrxItemProcessorContext processorCtx = executorService.createProcessorContext(ctx, trx);
 		return processorCtx;
 	}