+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/allocation/IAllocationDestination.java
@@ -46,5 +46,7 @@ public interface IAllocationDestination
 	 *
 	 * @param huContext
 	 */
-	void loadComplete(IHUContext huContext);
+	default void loadComplete(IHUContext huContext)
+	{
+	}
 }