+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/pporder/api/impl/HUPPCostCollectorBL.java
@@ -13,15 +13,14 @@ package de.metas.handlingunits.pporder.api.impl;
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
-
 
 import java.util.Collection;
 import java.util.List;
@@ -30,9 +29,15 @@ import org.adempiere.model.IContextAware;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.util.Check;
 import org.adempiere.util.Services;
+import org.eevolution.api.IPPCostCollectorBL;
+import org.eevolution.api.IReceiptCostCollectorCandidate;
+
+import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableList;
 
 import de.metas.handlingunits.IHUAssignmentBL;
 import de.metas.handlingunits.IHUAssignmentDAO;
+import de.metas.handlingunits.IHandlingUnitsBL;
 import de.metas.handlingunits.exceptions.HUException;
 import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_PP_Cost_Collector;
@@ -41,6 +46,30 @@ import de.metas.handlingunits.snapshot.IHUSnapshotDAO;
 
 public class HUPPCostCollectorBL implements IHUPPCostCollectorBL
 {
+	@Override
+	public I_PP_Cost_Collector createReceipt(final IReceiptCostCollectorCandidate candidate, final I_M_HU hu)
+	{
+		Preconditions.checkNotNull(candidate, "candidate is null");
+		Preconditions.checkNotNull(hu, "hu is null");
+
+		// services
+		final IPPCostCollectorBL ppCostCollectorBL = Services.get(IPPCostCollectorBL.class);
+		final IHandlingUnitsBL handlingUnitsBL = Services.get(IHandlingUnitsBL.class);
+
+		//
+		// Create & process the receipt cost collector
+		final I_PP_Cost_Collector cc = InterfaceWrapperHelper.create(ppCostCollectorBL.createReceipt(candidate), I_PP_Cost_Collector.class);
+
+		// Assign the HU to cost collector
+		assignHUs(cc, ImmutableList.of(hu));
+
+		//
+		// Activate the HU (assuming it was Planning)
+		handlingUnitsBL.setHUStatusActive(ImmutableList.of(hu));
+
+		return cc;
+	}
+
 	@Override
 	public void assignHUs(final org.eevolution.model.I_PP_Cost_Collector cc, final Collection<I_M_HU> husToAssign)
 	{