+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/invoicecandidate/process/C_Invoice_Candidate_GenerateInvoice.java
@@ -90,7 +90,10 @@ public class C_Invoice_Candidate_GenerateInvoice extends JavaProcess
 			final String notificationsLink = linkHelper.createShowWindowHTML(
 					"@AD_Note_ID@ #" + result.getNotificationCount(),
 					I_AD_Note.Table_Name,
-					result.getNotificationsWhereClause());
+					-1, // suggested windowID -> use the default one
+					result.getNotificationsWhereClause()
+					
+					);
 			summary.append(", ").append(notificationsLink);
 		}