+++ de.metas.swat/de.metas.swat.base/src/main/sql/postgresql/system/40-sw01_swat/5460140_sys_gh1064_docType_Rules_Changed.sql
@@ -1,0 +1,24 @@
+-- 2017-04-13T10:24:42.221
+-- URL zum Konzept
+UPDATE AD_Field SET DisplayLogic='@DocBaseType@=''SOO'' | @DocBaseType@=''POO'' | @DocBaseType@=''ARI'' | @DocBaseType@=''ARC'' | @DocBaseType@=''MOP'' | @DocBaseType@=''MMR'' | @DocBaseType@=''MMS'' | @DocBaseType@=''API'' | @DocBaseType@=''MMI''',Updated=TO_TIMESTAMP('2017-04-13 10:24:42','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Field_ID=2581
+;
+
+-- 2017-04-13T10:26:10.275
+-- URL zum Konzept
+UPDATE AD_Val_Rule SET Code=' (''@DocBaseType@''=''ARI'' AND AD_Ref_List.Value IN (''AQ'', ''AP''))
+ OR (''@DocBaseType@''=''ARC'' AND AD_Ref_List.Value IN (''CQ'', ''CR'',''CS''))
+ OR (''@DocBaseType@'' IN(''API'', ''MOP'') AND AD_Ref_List.Value IN (''QI'', ''VI''))
+OR (''@DocBaseType@'' = ''MMI'' AND AD_Ref_List.Value = ''MD'')
+ OR (''@DocBaseType@'' NOT IN (''API'', ''ARI'', ''ARC'', ''MOP'') AND AD_Ref_List.Value NOT IN (''AQ'', ''AP'', ''CQ'', ''CR'', ''QI'')) /* fallback for the rest of the entries */
+',Updated=TO_TIMESTAMP('2017-04-13 10:26:10','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Val_Rule_ID=540219
+;
+
+-- 2017-04-13T10:27:16.408
+-- URL zum Konzept
+INSERT INTO AD_Ref_List (AD_Client_ID,AD_Org_ID,AD_Ref_List_ID,AD_Reference_ID,Created,CreatedBy,EntityType,IsActive,Name,Updated,UpdatedBy,Value,ValueName) VALUES (0,0,541260,148,TO_TIMESTAMP('2017-04-13 10:27:16','YYYY-MM-DD HH24:MI:SS'),100,'de.metas.swat','Y','Material Disposal',TO_TIMESTAMP('2017-04-13 10:27:16','YYYY-MM-DD HH24:MI:SS'),100,'MD','Material Disposal')
+;
+
+-- 2017-04-13T10:27:16.420
+-- URL zum Konzept
+INSERT INTO AD_Ref_List_Trl (AD_Language,AD_Ref_List_ID, Description,Name, IsTranslated,AD_Client_ID,AD_Org_ID,Created,Createdby,Updated,UpdatedBy) SELECT l.AD_Language,t.AD_Ref_List_ID, t.Description,t.Name, 'N',t.AD_Client_ID,t.AD_Org_ID,t.Created,t.Createdby,t.Updated,t.UpdatedBy FROM AD_Language l, AD_Ref_List t WHERE l.IsActive='Y' AND l.IsSystemLanguage='Y' AND l.IsBaseLanguage='N' AND t.AD_Ref_List_ID=541260 AND NOT EXISTS (SELECT * FROM AD_Ref_List_Trl tt WHERE tt.AD_Language=l.AD_Language AND tt.AD_Ref_List_ID=t.AD_Ref_List_ID)
+;