+++ ReleaseNotes.md
@@ -10,17 +10,31 @@ Additional notes:
 
 Here come the actual release notes:
 
-# metasfresh 5.5 (2017-16)
+# metasfresh 5.6 (2017-16)
 
 **upcoming release for week 2017-16**
 
 ## Features
+
+* metasfresh-backend
+  * [#1062](https://github.com/metasfresh/metasfresh/issues/1062) Return HU to vendor
+  * [#1064](https://github.com/metasfresh/metasfresh/issues/1064) Move HU to garbage
+  * [#1065](https://github.com/metasfresh/metasfresh/issues/1065) Move HU to other Warehouse/ Locator
+  * [#1286](https://github.com/metasfresh/metasfresh/issues/1286) Performance Business Partner Pricelist generation process
+
+* metasfresh-webui-api
+  * [#269](https://github.com/metasfresh/metasfresh-webui-api/issues/269) API for Manufacturing UI Planning/ Doing
+  
 * metasfresh-frontend-api
   * [#608](https://github.com/metasfresh/metasfresh-webui-frontend/issues/608) Chart animations
     * Adding animation to dashboard charts (pie and bar charts).
 
 ## Fixes
-
+* metasfresh-backend
+  * [#1191](https://github.com/metasfresh/metasfresh/issues/1191) small adjustments in jasper documents
+    * fix empty page in sales invoice document
+  * [#1264](https://github.com/metasfresh/metasfresh/issues/1264) Material Receipt document best before date missing
+    
 # metasfresh 5.5 (2017-15)
 
 ## Features